Fingerprint set recovery broken.
This implementation used to group n fingerprints into a sorted
in-memory page. Pages were subsequently merged on-disk directly,
creating the on-disk storage file for DiskFPSet.

The new algorithm simply "replays" the fingerprints found in the
trace file. It's biggest disadvantage is a performance penalty it
pays because it doesn't group fingerprints. On the other hand, it has
advantages over the old algorithm:

- Simplified logic/code
- No need for a long[] recovery buffer
- TLC runs with a warm in-memory fingerprint cache
- With large amounts of available fingerprint set memory, the .fp
file might actually never be written. This means that the FPSet never
has to go to disk during contains/put which yields a better overall
runtime performance.


[Bug][TLC]